servers: changed size of syscall ID and adjusted alignment

+++ src/include/kernel/servers.h
@@ -61,7 +61,7 @@ typedef struct {
 
 typedef struct {
     MessageHeader header;
-    uint64_t id;            // syscall request ID
+    uint16_t id;            // syscall request ID
 } SyscallHeader;
 
 /* log command */
@@ -137,7 +137,7 @@ typedef struct {
     gid_t gid;
     off_t position;
     size_t length;
-    uint8_t data[];
+    uint64_t data[];    // for alignment
 } RWCommand;
 
 /* IRQ Notification */
